requirements.txt
# Testing
mock==1.0.1
coverage==3.7.1
django-nose==1.2
coveralls==0.4.2
django-webtest==1.7.7
tox==1.7.1
detox==0.9.3
WebTest==2.0.15
pinocchio==0.4.1
requests==2.3.0

# Development
django-extensions==1.3.3
django-debug-toolbar==1.2
flake8==2.1.0
